remove process management and fix up logging
The logic is much simpler to just avoid trying to bundle in a process
manager with the main code path. You end up needing to manage daemons
that require bookkeeping such as pidfiles and specific OS
support. This is out of scope for xenv as I'd like to work on Windows.
The other option is to keep the processes in the same process tree,
but it becomes really difficult to manage relationship between the
"main" process and "service" processes along with what might need to
be restarted based on the configuration.

My solution then is to punt! Instead, if you need a service to restart
when configuration changes, xenv will kill it for you and expecte a
process manager such as supervisord, systemd or k8s to restart things
accordingly to the platform.

I'm leaving the manager code in case it is beneficial to create a
simple daemon-like manager that starts processes and keeps track of
some bookkeeping for local development.